### 2025-10-31 #9329 @editorialbot
[REVIEW]: graph-pes: graph-based machine-learning models for potential-energy surfaces
**Submitting author:** @jla-gardner (<a href="http://orcid.org/0009-0006-7377-7146">John Gardner</a>)
**Repository:** https://github.com/vldgroup/graph-pes
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.0.0
