### 2025-12-06 #9329 @HectorMozo3110
Hi, @jla-gardner I have reviewed the paper again and, regarding the word limit, it seems appropriate. I only have one observation regarding the Statement of Need section: although the section introduces the general context well and explains who the software is intended for, it still lacks a more explicit justification of the need for the package. It would be important to include a direct comparison with similar existing tools, explain the specific limitations those alternatives have, and clarify what graph-pes provides that truly fills a gap within the community. In other words, while the diversity of MLIPs and the lack of common interfaces are mentioned, it would be helpful to expand on why this is a practical issue and how graph-pes uniquely addresses it. Everything else looks good and the manuscript is solid. I have been testing the installation and other related aspects; I will update those points over the weekend.
