+++ paper/paper.md
@@ -32,14 +32,14 @@ We present `graph-pes`, an open-source toolkit for accelerating the development,
  
 2. **The `graph-pes-train` command-line interface** (CLI): a convenience tool for training graph-based MLIPs on datasets of labelled atomic structures directly from the command line. The tool is compatible with any `GraphPESModel` (i.e., those defined in `graph-pes`, user-designed ones, and foundation models) and is designed to be easily extensible via custom loss functions, optimisers, datasets, and more.
 
-3. **Molecular-dynamics drivers** that allow any `GraphPESModel` to be used in GPU-accelerated MD simulations. We currently provide a `pair style` for LAMMPS [@Thompson-22-02], a `GraphPESCalculator` for ASE [@Larsen-17-06], and an integration with the `torch-sim` package [@torch-sim].
+3. **Molecular-dynamics drivers** that allow any `GraphPESModel` to be used in GPU-accelerated MD simulations. We currently provide a `pair style` for LAMMPS [@Thompson-22-02], a `GraphPESCalculator` for `ASE` [@Larsen-17-06], and an integration with the `torch-sim` package [@torch-sim].
 
 # Statement of need
 
 In recent years, machine-learned PES models have become central tools for computational chemistry and materials science [@Deringer-19-11].
 These models are trained on labels generated by quantum-mechanical methods, but scale much more favourably with system size, making it possible to simulate the dynamics of large systems over extended timescales.
 
-Many flavours of MLIPs exist, and with them have arisen a variety of software packages, each typically tailored to training specific architectures. Given their unique specialisations, these individual implementations do not normally conform to a common interface, making it difficult for practitioners to migrate training and validation pipelines between different architectures.
+Many flavours of MLIPs exist, and with them have arisen a variety of software packages, each typically tailored to training specific architectures (see below). Given their unique specialisations, these individual implementations do not normally conform to a common interface, making it difficult for practitioners to migrate training and validation pipelines between different architectures.
 
 ![Schematic overview of the functionality of `graph-pes`. The core components are highlighted in colour. Red: The `AtomicGraph` class is used to represent atomic structures and incorporates the notion of locality via a neighbour list. Blue: The `GraphPESModel` class is the general base class for all `graph-pes` models. Green: `graph-pes` includes a CLI for easy training, and interfaces to multiple external simulation tools for evaluating MLIPs.](./overview.png)
 
@@ -49,7 +49,10 @@ Many flavours of MLIPs exist, and with them have arisen a variety of software pa
 
 `graph-pes` is beginning to drive projects within our research group, and we hope that it will be useful to many others. In recent work, we have described the use of `graph-pes` for fitting NequIP models to datasets created using `autoplex` [@Liu-25-08], for assessing zero-shot performance of different graph-network MLIPs [@Mahmoud-25-02], and for distilling atomistic foundation models [@Gardner-25-06].
 
-Relevant alternative packages that offer training and validation for _specific_ ML-PES architectures include: `schnetpack` [@schutt2019schnetpack; @schutt2023schnetpack], `deepmd-kit` [@Wang-18-07; @Zeng-23-08], `nequip` [@Batzner-22-05], `mace-torch` [@Batatia-22-10], `torchmd-net` [@TorchMDNet], and `fairchem` [@fairchem]. The `MatterTune` package [@Kong-25-04] provides a unified interface for fine-tuning atomistic foundation models, but does not create models with a common interface, or allow for training arbitrary MLIP architectures from scratch.
[…]
