+++ paper/paper.md
@@ -43,7 +43,7 @@ Many flavours of MLIPs exist, and with them have arisen a variety of software pa
 
 ![Schematic overview of the functionality of `graph-pes`. The core components are highlighted in colour. Red: The `AtomicGraph` class is used to represent atomic structures and incorporates the notion of locality via a neighbour list. Blue: The `GraphPESModel` class is the general base class for all `graph-pes` models. Green: `graph-pes` includes a CLI for easy training, and interfaces to multiple external simulation tools for evaluating MLIPs.](./overview.png)
 
-`graph-pes` provides a **unified interface and framework** for defining, training, and working with graph-based MLIP models. This reduces the barrier to entry for researchers wanting to implement new MLIP architectures, and allows practitioners to easily explore different MLIP architectures: training scripts require as little as one line of code to swap between model architectures, while validation scripts can be written in an architecture-agnostic manner, with `LAMMPS` input scripts, `ASE` calculators, and `torch-sim` simulations requiring no changes other than pointing to a different model file.
+`graph-pes` provides a **unified interface and framework** for defining, training, and working with graph-based MLIP models, complementing existing software in the field (see "Related work" section below). This reduces the barrier to entry for researchers wanting to implement new MLIP architectures, and allows practitioners to easily explore different MLIP architectures: training scripts require as little as one line of code to swap between model architectures, while validation scripts can be written in an architecture-agnostic manner, with `LAMMPS` input scripts, `ASE` calculators, and `torch-sim` simulations requiring no changes other than pointing to a different model file.
 
 # Related work
 
