### 2025-12-19 #9329 @jla-gardner
Hi @HectorMozo3110 , sorry for the delay, and thank you for the suggestion.

We have revised our Statement of Need section, as requested, stating briefly how graph-pes compares to existing packages.

Furthermore, in the Related Work section, we now draw a clearer comparison between graph-pes and other existing packages to be more explicit, and added relevant references:

> A number of existing packages offer training and validation pipelines for particular ML-PES architectures, including `schnetpack`, `deepmd-kit`, `nequip`, `mace-torch`, `torchmd-net`, and `fairchem`. These frameworks focus on their associated model families and do not share a common interface for training. While `MatterTune` offers a unified interface for foundation model fine-tuning, it does not easily support training arbitrary models from scratch. In contrast to these, `graph-pes` is a general, model-agnostic framework, designed to enable exact side-by-side comparisons, easy implementation of arbitrary new architectures, and standardized training and evaluation workflows.

To avoid duplication, we have kept this explicit discussion within the Related Work section, but reference it suitably in the Statement of Need in an attempt to tie these sections together. I hope this is a suitable addition that resolves the remaining change request?
