+++ README.md
@@ -1,17 +1,17 @@
 <div align="center">
-    <a href="https://jla-gardner.github.io/graph-pes/">
+    <a href="https://vldgroup.github.io/graph-pes/">
         <img src="docs/source/_static/logo-text.svg" width="90%"/>
     </a>
 
 `graph-pes` is a framework built to accelerate the development of machine-learned potential energy surface (PES) models that act on graph representations of atomic structures.
 
-Links: [Google Colab Quickstart](https://colab.research.google.com/github/jla-gardner/graph-pes/blob/main/docs/source/quickstart/quickstart.ipynb) - [Documentation](https://jla-gardner.github.io/graph-pes/) - [PyPI](https://pypi.org/project/graph-pes/)
+Links: [Google Colab Quickstart](https://colab.research.google.com/github/vldgroup/graph-pes/blob/main/docs/source/quickstart/quickstart.ipynb) - [Documentation](https://vldgroup.github.io/graph-pes/) - [PyPI](https://pypi.org/project/graph-pes/)
 
 [![PyPI](https://img.shields.io/pypi/v/graph-pes)](https://pypi.org/project/graph-pes/)
 [![Conda-forge](https://img.shields.io/conda/vn/conda-forge/graph-pes.svg)](https://github.com/conda-forge/graph-pes-feedstock)
-[![Tests](https://github.com/jla-gardner/graph-pes/actions/workflows/tests.yaml/badge.svg?branch=main)](https://github.com/jla-gardner/graph-pes/actions/workflows/tests.yaml)
-[![codecov](https://codecov.io/gh/jla-gardner/graph-pes/branch/main/graph/badge.svg)](https://codecov.io/gh/jla-gardner/graph-pes)
-[![GitHub last commit](https://img.shields.io/github/last-commit/jla-gardner/load-atoms)]()
+[![Tests](https://github.com/vldgroup/graph-pes/actions/workflows/tests.yaml/badge.svg?branch=main)](https://github.com/vldgroup/graph-pes/actions/workflows/tests.yaml)
+[![codecov](https://codecov.io/gh/vldgroup/graph-pes/branch/main/graph/badge.svg)](https://codecov.io/gh/vldgroup/graph-pes)
+[![GitHub last commit](https://img.shields.io/github/last-commit/vldgroup/graph-pes)]()
 
 </div>
 
@@ -26,12 +26,12 @@ As a researcher wanting to work on MLIP methodology, `graph-pes` makes implement
 
 ## Features
 
-- Experiment with new model architectures by inheriting from our `GraphPESModel` [base class](https://jla-gardner.github.io/graph-pes/models/root.html).
-- [Train your own](https://jla-gardner.github.io/graph-pes/quickstart/implement-a-model.html) or existing model architectures (e.g., [SchNet](https://jla-gardner.github.io/graph-pes/models/many-body/schnet.html), [NequIP](https://jla-gardner.github.io/graph-pes/models/many-body/nequip.html), [PaiNN](https://jla-gardner.github.io/graph-pes/models/many-body/pinn.html), [MACE](https://jla-gardner.github.io/graph-pes/models/many-body/mace.html), [TensorNet](https://jla-gardner.github.io/graph-pes/models/many-body/tensornet.html), [OrB](https://jla-gardner.github.io/graph-pes/models/many-body/orb.html) etc.).
[…]
