+++ _sources/development.rst.txt
@@ -1,0 +1,5 @@
+Development
+===========
+
+We welcome any suggestions and contributions to this project.
+Please visit our `GitHub repository <https://github.com/jla-gardner/graph-pes>`_ to report issues or submit pull requests.
