### 2026-02-20 #9329 @jla-gardner
@yuanqing-wang thank you for this! The release is on Zenodo at https://zenodo.org/records/18714766. The version is v1.0.0
