### 2026-03-04 #9329 @jla-gardner
Hi @yuanqing-wang 👋

We’ve noticed some trivial updates to make to the manuscript before submission:
1. two of our references have been released as papers (as opposed to the existing pre-prints) so we’d like to bump those
2. the location of this repo has now changed to https://github.com/vldgroup/graph-pes, but the "Repository" link still links to https://github.com/jla-gardner/graph-pes

Are you able to update the link of this repo via the editorialbot? If we then push these trivial changes, are you happy to proceed? What are the next steps if so?
