+++ paper/paper.bib
@@ -190,17 +190,16 @@
   urldate = {2022-11-08}
 }
 
-@misc{Kovacs-25-01,
-  title         = {{{MACE-OFF}}: {{Transferable Short Range Machine Learning Force Fields}} for {{Organic Molecules}}},
-  author        = {Kov{\'a}cs, D{\'a}vid P{\'e}ter and Moore, J. Harry and Browning, Nicholas J. and Batatia, Ilyes and Horton, Joshua T. and Pu, Yixuan and Kapil, Venkat and Witt, William C. and Magd{\u a}u, Ioan-Bogdan and Cole, Daniel J. and Cs{\'a}nyi, G{\'a}bor},
-  year          = {2025},
-  month         = jan,
-  number        = {arXiv:2312.15211},
-  eprint        = {2312.15211},
-  primaryclass  = {physics},
-  doi           = {10.48550/arXiv.2312.15211},
-  urldate       = {2025-01-30},
-  archiveprefix = {arXiv}
+@article{Kovacs-25-05,
+  title     = {{{MACE-OFF}}: {{Short-Range Transferable Machine Learning Force Fields}} for {{Organic Molecules}}},
+  author    = {Kov{\'a}cs, D{\'a}vid P{\'e}ter and Moore, J. Harry and Browning, Nicholas J. and Batatia, Ilyes and Horton, Joshua T. and Pu, Yixuan and Kapil, Venkat and Witt, William C. and Magd{\u a}u, Ioan-Bogdan and Cole, Daniel J. and Cs{\'a}nyi, G{\'a}bor},
+  year      = 2025,
+  journal   = {Journal of the American Chemical Society},
+  volume    = {147},
+  number    = {21},
+  pages     = {17598--17611},
+  doi       = {10.1021/jacs.4c07099},
+  copyright = {https://creativecommons.org/licenses/by/4.0/}
 }
 
 @article{Batatia-25-11,
@@ -279,18 +278,17 @@
   archiveprefix = {arXiv}
 }
 
-@misc{Mahmoud-25-02,
-  title         = {Assessing Zero-Shot Generalisation Behaviour in Graph-Neural-Network Interatomic Potentials},
-  author        = {Ben Mahmoud, Chiheb and {El-Machachi}, Zakariya and Gierczak, Krystian A. and Gardner, John L. A. and Deringer, Volker L.},
-  year          = {2025},
-  month         = feb,
-  number        = {arXiv:2502.21317},
-  eprint        = {2502.21317},
-  primaryclass  = {physics},
-  doi           = {10.48550/arXiv.2502.21317},
-  urldate       = {2025-03-04},
-  archiveprefix = {arXiv}
+@article{Mahmoud-25-11,
+  title   = {Assessing Zero-Shot Generalisation Behaviour in Graph-Neural-Network Interatomic Potentials},
+  author  = {Mahmoud, Chiheb Ben and {El-Machachi}, Zakariya and Gierczak, Krystian A. and Gardner, John L. A. and Deringer, Volker L.},
+  year    = 2025,
+  journal = {Digital Discovery},
+  volume  = {4},
+  number  = {11},
+  pages   = {3389--3399},
+  doi     = {10.1039/D5DD00103J}
 }
+
 
 @article{Pickard-22-07,
   title   = {Ephemeral Data Derived Potentials for Random Structure Search},
