+++ paper/paper.bib
@@ -280,7 +280,7 @@
 
 @article{Mahmoud-25-11,
   title   = {Assessing Zero-Shot Generalisation Behaviour in Graph-Neural-Network Interatomic Potentials},
-  author  = {Mahmoud, Chiheb Ben and {El-Machachi}, Zakariya and Gierczak, Krystian A. and Gardner, John L. A. and Deringer, Volker L.},
+  author  = {Ben Mahmoud, Chiheb and {El-Machachi}, Zakariya and Gierczak, Krystian A. and Gardner, John L. A. and Deringer, Volker L.},
   year    = 2025,
   journal = {Digital Discovery},
   volume  = {4},
