+++ paper/paper.bib
@@ -39,13 +39,16 @@
 }
 
 @inproceedings{Batatia-22-10,
-  title     = {{{MACE}}: {{Higher Order Equivariant Message Passing Neural Networks}} for {{Fast}} and {{Accurate Force Fields}}},
-  booktitle = {Advances in {{Neural Information Processing Systems}}},
-  author    = {Batatia, Ilyes and Kovacs, David Peter and Simm, Gregor N. C. and Ortner, Christoph and Csanyi, Gabor},
-  year      = {2022},
-  month     = oct,
-  urldate   = {2025-06-22}
+  author    = {Batatia, Ilyes and Kovacs, David P and Simm, Gregor and Ortner, Christoph and Csanyi, Gabor},
+  booktitle = {Advances in Neural Information Processing Systems},
+  editor    = {S. Koyejo and S. Mohamed and A. Agarwal and D. Belgrave and K. Cho and A. Oh},
+  pages     = {11423--11436},
+  title     = {MACE: Higher Order Equivariant Message Passing Neural Networks for Fast and Accurate Force Fields},
+  url       = {https://proceedings.neurips.cc/paper_files/paper/2022/file/4a36c3c51af11ed9f34615b81edb5bbc-Paper-Conference.pdf},
+  volume    = {35},
+  year      = {2022}
 }
+
 
 
 @article{HjorthLarsen-17-06,
@@ -70,26 +73,21 @@
   urldate = {2022-11-08}
 }
 
-
-@misc{Simeon-23-06,
-  title         = {{{TensorNet}}: {{Cartesian Tensor Representations}} for {{Efficient Learning}} of {{Molecular Potentials}}},
-  author        = {Simeon, Guillem and {de Fabritiis}, Gianni},
-  year          = {2023},
-  number        = {arXiv:2306.06482},
-  eprint        = {2306.06482},
-  primaryclass  = {physics},
-  archiveprefix = {arXiv}
+@inproceedings{Simeon-23-06,
+  title     = {TensorNet: Cartesian Tensor Representations for Efficient Learning of Molecular Potentials},
+  author    = {Simeon, Guillem and {de Fabritiis}, Gianni},
+  booktitle = {Thirty-seventh Conference on Neural Information Processing Systems},
+  year      = {2023},
+  url       = {https://openreview.net/forum?id=BEHlPdBZ2e}
 }
 
-@misc{Schutt-21-06,
-  title         = {Equivariant Message Passing for the Prediction of Tensorial Properties and Molecular Spectra},
-  author        = {Sch{\"u}tt, Kristof T. and Unke, Oliver T. and Gastegger, Michael},
-  year          = {2021},
-  number        = {arXiv:2102.03150},
-  eprint        = {2102.03150},
-  primaryclass  = {physics},
-  doi           = {10.48550/arXiv.2102.03150},
-  archiveprefix = {arXiv}
+@inproceedings{Schutt-21-06,
+  title     = {Equivariant Message Passing for the Prediction of Tensorial Properties and Molecular Spectra},
+  author    = {Sch{\"u}tt, Kristof T. and Unke, Oliver T. and Gastegger, Michael},
+  booktitle = {Proceedings of Machine Learning Research},
+  year      = {2021},
+  note      = {Pre-print at: \url{https://arxiv.org/abs/2102.03150}},
+  doi       = {10.48550/arXiv.2102.03150}
 }
 
 @article{Wang-18-07,
@@ -352,17 +350,15 @@
   year    = {2025}
 }
 
-@misc{Kong-25-04,
[…]
