+++ paper/paper.bib
@@ -41,15 +41,11 @@
 @inproceedings{Batatia-22-10,
   author    = {Batatia, Ilyes and Kovacs, David P and Simm, Gregor and Ortner, Christoph and Csanyi, Gabor},
   booktitle = {Advances in Neural Information Processing Systems},
-  editor    = {S. Koyejo and S. Mohamed and A. Agarwal and D. Belgrave and K. Cho and A. Oh},
   pages     = {11423--11436},
   title     = {MACE: Higher Order Equivariant Message Passing Neural Networks for Fast and Accurate Force Fields},
-  url       = {https://proceedings.neurips.cc/paper_files/paper/2022/file/4a36c3c51af11ed9f34615b81edb5bbc-Paper-Conference.pdf},
   volume    = {35},
   year      = {2022}
 }
-
-
 
 @article{HjorthLarsen-17-06,
   title   = {The Atomic Simulation Environment---a {{Python}} Library for Working with Atoms},
@@ -78,7 +74,7 @@
   author    = {Simeon, Guillem and {de Fabritiis}, Gianni},
   booktitle = {Thirty-seventh Conference on Neural Information Processing Systems},
   year      = {2023},
-  url       = {https://openreview.net/forum?id=BEHlPdBZ2e}
+  doi       = {10.48550/arXiv.2306.06482}
 }
 
 @inproceedings{Schutt-21-06,
@@ -86,7 +82,6 @@
   author    = {Sch{\"u}tt, Kristof T. and Unke, Oliver T. and Gastegger, Michael},
   booktitle = {Proceedings of Machine Learning Research},
   year      = {2021},
-  note      = {Pre-print at: \url{https://arxiv.org/abs/2102.03150}},
   doi       = {10.48550/arXiv.2102.03150}
 }
 
