+++ paper/paper.bib
@@ -45,6 +45,7 @@
   title     = {MACE: Higher Order Equivariant Message Passing Neural Networks for Fast and Accurate Force Fields},
   volume    = {35},
   year      = {2022}
+  doi       = {10.48550/arXiv.2206.07697}
 }
 
 @article{HjorthLarsen-17-06,
