+++ paper/paper.bib
@@ -44,7 +44,7 @@
   pages     = {11423--11436},
   title     = {MACE: Higher Order Equivariant Message Passing Neural Networks for Fast and Accurate Force Fields},
   volume    = {35},
-  year      = {2022}
+  year      = {2022},
   doi       = {10.48550/arXiv.2206.07697}
 }
 
