### 2026-04-06 cad1d5a
Polish TorchSim GraphPES integration
