### 2026-09-03 #9329 @abhishektiwari
@iammix @vwxyzjn would any of you be willing to review this submission for Journal of Open Source Software (JOSS)? The submission under consideration is "graph-pes: graph-based machine-learning models for potential-energy surfaces".

Based on your past review of related JOSS publications, we think you might be able to provide a great review of this submission.

Thanks
Abhi
